+++ PharmaUI/ReceiptPayment/frmReceiptFromCustomer.cs
@@ -210,7 +210,7 @@ namespace PharmaUI.ReceiptPayment
 
             //Display totall of adjusted amount
             decimal totallAdjusted = 0;
-            allAdjustment.ForEach(x => totallAdjusted += x.Amount);
+            allAdjustment.ForEach(x => totallAdjusted += (decimal)x.Amount);
             lblAmtAdjVal.Text = Convert.ToString(totallAdjusted);
         }
